Make a start on the Agent part of the application
The Agent is the smart part, so will take an instance of the game tree and search it for the optimal move.
Once the optimal move is found, the Agent will make that move and wait for its turn to come round again.

Remove some of the coupling from min_max, the plan is to make min_max more modular and generic.